+++ pqs/__init__.py
@@ -5,9 +5,11 @@ Submodules:
 """
 # Libraries
 # # Internal
-from .start import Base, SessionFactory, bind_engine
+from .start import Base, SessionFactory, bind_engine, is_engine_binded, \
+                   get_engine
 from .binder import PQSFieldBinder
 from .binders_sample import QLineEditBinder, QDateEditBinder
+from .connect import PQSConnectUI, PQSConnectDefaultUI
 from .editor import PQSEditUI
 from .errors import PQSUINotValidError, PQSModelNotValidError, \
                     PQSSessionNotValidError, PQSFormFieldsNotValidError, \
@@ -18,4 +20,6 @@ __all__ = ["Base", "SessionFactory", "bind_engine", "PQSFieldBinder",
            "QLineEditBinder", "QDateEditBinder", "PQSEditUI",
            "PQSUINotValidError", "PQSSessionNotValidError",
            "PQSModelNotValidError", "PQSFormFieldsNotValidError",
-           "PQSBindersNotFoundError", "PQSBindersNotValidError"]
+           "PQSBindersNotFoundError", "PQSBindersNotValidError",
+           "PQSConnectDefaultUI", "PQSConnectUI", "is_engine_binded",
+           "get_engine"]